dashboard/requirements.txt
python-jobspy>=1.1.0
requests>=2.28.0
